Move non human variant calling pipeline steps into qc and variant calling pipelines gatk variant calling pipeline to output vcfstats file
